Added interface for a basic object which can have states
