+++ src/commonMain/kotlin/fsm/StateManager.kt
@@ -6,44 +6,6 @@ import kotlin.reflect.KClass
  * Create a new [StateMachine] and optionally apply a callback
  */
 
-interface Stateable {
-    val manager: StateManager
-
-    fun setStartState(state: StateExecutor) {
-        manager.setStartState(state)
-    }
-
-    fun doStateChange(state: StateExecutor) {
-        manager.doStateChange(state)
-    }
-
-    fun updateCurrentState() {
-        manager.updateCurrentState()
-    }
-
-    fun getCurrentState(): StateExecutor {
-        return manager.getCurrentState()
-    }
-}
-
-inline fun <reified T: Stateable> T.declareState(): StateExecutor {
-    val currentManager = StateManager.stateMachines[T::class]
-    if (currentManager != null) {
-        return currentManager.createState()
-    }
-    error("You have to declare a manager first!")
-}
-
-inline fun <reified T: Stateable> T.useStates(callback: StateManager.() -> Unit = {}): StateManager {
-    val manager = StateManager()
-    StateManager.stateMachines[T::class] = manager
-    manager.apply { callback() }
-    return manager
-}
-
-/**
- * This class manages all the states of a class-object(for example [Entity]).
- */
 class StateManager {
 
     companion object {
@@ -71,8 +33,12 @@ class StateManager {
     fun setStartState(state: StateExecutor) = doStateChange(state)
 
     //creates a new fsm.old.State for the object called "owner" -> the state only affects him
-    fun createState(): StateExecutor {
-        val stateBase = State()
+    fun createState(onBegin: () -> Unit = {}, onExecute: () -> Unit = {}, onEnd: () -> Unit = {}): StateExecutor {
+        val stateBase = State().apply {
+            begin = onBegin
+            execute = onExecute
+            end = onEnd
+        }
         val state = StateExecutor(stateBase)
         this.states[stateBase.id] = state
         return state
